Removed jellyfish version... we now link to jellyfish externally via "--with-jellyfish-dir=<PATH>" when running the configure command.  This should find all the relevant libraries and header files... note we are fixed with 1.1.10 for the time being.
Added functionality to sect.  We can now produce matricies from the GC% and CVG information which can be used to detect contamination in the assemblies.

Added feature in flame plot which automatically determines size of matrix to plot.

Added architecture for unit testing... just need to find a decent framework to incorperate which isn't too heavy.