Modified configure script, so handling dependencies is easier.
Removed kseq files (should ahve been in previous commit)

Added a QT project file, to make it easier to work with the source code.  There is no requirement for other developers to use this but I've added it to make it easier for my to work with the project if I move it to another directory.