Removed call to auto configure when running the autogen.sh script.
Fixed issue where title could not be overriden by the user in the plotting
tools.

Changed short option name for the "list" option in spectra-mx plot tool to
'c'.  This is to avoid a conflict with the "title" option.

Provided a more descriptive error message if the user enters a malformed
value for the "list" option in the spectra-mx tool.

Removed call to set the size ratio to 1 in the spectra-mx tool