Updated version to 1.0.6.
Added check for strerror in configure.ac.  This allows compatibility with jellyfish 1.1.11.

Fixed a bug with the spectra-cn plots where we duplicated the entry for the highest copy number displayed before combining all together.

Added a feature to spectra-cn so that we can show cumulative increases in distinct kmers as kmer coverage increases.