Fix broken optionality of input_3 in comp.cc
Leaving out the optional input_3 (the filter) was causing an error to be
thrown.

To make it optional again added quick check to comp.cc to not run
input_handling if input_3 is an empty string.

Preserves fixed error handling for missing/incorrect input_1/input_2
and correctly throws error when non-existent file is provided as input_3