Updating version to 2.2.0.
Main change here is forcing static linking of kat_jellyfish and kat libraries into the kat executable.  Before we were using libtool to dynamically link the libraries, which doesn't make much sense for the kat executable.  This way is more robust and more likely to deliver expected behaviour, particularly when developing the codebase, as there is less chance of linking to an unexpected version of the library.